requirements.txt
# requirements.txt
kafka-python==2.0.2
# confluent-kafka[avro]==2.3.0
protobuf
requests==2.31.0
psycopg2-binary==2.9.7
pyspark==3.4.1
pandas==2.0.3
numpy
plotly==5.17.0
dash==2.14.1
python-dotenv==1.0.0
schedule==1.2.0